+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Change Log
-- Use flag `(?i)` to color any words no matter lowercase and uppercase
-- New `diagnosticProvider.ts` fix to work with flag
-- Up to version 0.1.7
+- Add `Quotes detector` for `diagnosticProvider.ts`
+- Change `LICENSE` in `package.json`
+- Up to version 0.1.8